# DDS: speed up reading of uncompressed images